Move ClipCache out from ClipPlayer.java and into its own class file and remove thee "m_" prefix from variables
