+++ src/games/strategy/sound/ClipPlayer.java
@@ -588,47 +588,3 @@ public class ClipPlayer {
 }
 
 
-class ClipCache {
-  private final HashMap<URL, Clip> m_clipMap = new HashMap<URL, Clip>();
-  private final List<URL> m_cacheOrder = new ArrayList<URL>();
-  private final int MAXSIZE;
-
-  ClipCache(final int max) {
-    if (max < 1) {
-      throw new IllegalArgumentException("ClipCache max must be at least 1");
-    }
-    MAXSIZE = max;
-  }
-
-  public synchronized Clip get(final URL file) {
-    Clip clip = m_clipMap.get(file);
-    if (clip != null) {
-      m_cacheOrder.remove(file);
-      m_cacheOrder.add(file);
-      return clip;
-    }
-    if (m_clipMap.size() >= MAXSIZE) {
-      final URL leastPlayed = m_cacheOrder.get(0);
-      // System.out.println("Removing " + leastPlayed + " and adding " + file);
-      final Clip leastClip = m_clipMap.remove(leastPlayed);
-      leastClip.stop();
-      leastClip.flush();
-      leastClip.close();
-      m_cacheOrder.remove(leastPlayed);
-    }
-    clip = ClipPlayer.createClip(file, false);
-    m_clipMap.put(file, clip);
-    m_cacheOrder.add(file);
-    return clip;
-  }
-
-  public synchronized void removeAll() {
-    for (final Clip clip : m_clipMap.values()) {
-      clip.stop();
-      clip.flush();
-      clip.close();
-    }
-    m_clipMap.clear();
-    m_cacheOrder.clear();
-  }
-}
